refactor: less reapeated code
